+++ src/js/records/chain.record.js
@@ -6,6 +6,12 @@ export class ChainRecord {
     this.name = record.name
     this.bridgeContract = record.bridgeContract
     this.isBase = record.isBase
+
+    this.hexId = record.hexChainId
+    this.rpcUrl = record.rpc
+    this.blockExplorerUrl = record.explorer
+    this.nativeCurrencyName = record.nativeCurrencyName
+    this.nativeSymbol = record.nativeSymbol
     this.nativeDecimals = record?.nativeDecimals
   }
 }